file compiler warning: add cast for warning on OS X
git-svn-id: https://www.dnssec-tools.org/svn/dnssec-tools/trunk@8172 c4029f9c-b713-0410-a448-fae9e31a314a

+++ dnssec-tools/validator/apps/validator_driver.c
@@ -193,8 +193,8 @@ check_results(val_context_t * context, const char *desc, char * name,
 
     }
 
-    fprintf(stderr, "%s: ****END**** %ld.%ld sec\n", desc, duration.tv_sec,
-            duration.tv_usec);
+    fprintf(stderr, "%s: ****END**** %ld.%ld sec\n", desc,
+            (long) duration.tv_sec, (long) duration.tv_usec);
 
     return err;
 }
